+++ blockchain/wire.go
@@ -3,6 +3,7 @@ package blockchain
 import (
 	"github.com/tendermint/go-amino"
 	cryptoAmino "github.com/tendermint/tendermint/crypto/encoding/amino"
+	"github.com/tendermint/tendermint/types"
 )
 
 var cdc = amino.NewCodec()
@@ -10,4 +11,5 @@ var cdc = amino.NewCodec()
 func init() {
 	RegisterBlockchainMessages(cdc)
 	cryptoAmino.RegisterAmino(cdc)
+	types.RegisterEvidences(cdc)
 }