+++ consensus/types/wire.go
@@ -3,10 +3,12 @@ package types
 import (
 	"github.com/tendermint/go-amino"
 	cryptoAmino "github.com/tendermint/tendermint/crypto/encoding/amino"
+	"github.com/tendermint/tendermint/types"
 )
 
 var cdc = amino.NewCodec()
 
 func init() {
 	cryptoAmino.RegisterAmino(cdc)
+	types.RegisterEvidences(cdc)
 }